% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_gsea_results.R
\name{combine_gsea_results}
\alias{combine_gsea_results}
\title{Combine GSEA Results for All Samples}
\usage{
combine_gsea_results(gsea_results_list)
}
\arguments{
\item{gsea_results_list}{A list where each element is a GSEA result for a sample.}
}
\value{
A data frame containing GSEA results for all samples, with an added \code{sample} column to identify the source of each result.
}
\description{
This function takes the list of GSEA results for individual samples and combines them into a single data frame.
}
\examples{
combined_results <- combine_gsea_results(gsea_results_list)
}
