% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_score_data.R
\name{mp_score_data}
\alias{mp_score_data}
\title{Process GSEA Results and Calculate MP Scores}
\usage{
mp_score_data(gsea_final)
}
\arguments{
\item{gsea_final}{A data frame containing the combined GSEA results for all samples.}
}
\value{
A list with two elements: \code{gsea_res_clean} (the cleaned GSEA results) and \code{score_sum_by_cluster} (a summary of MP scores by sample).
}
\description{
This function processes the GSEA results and calculates MP scores for each sample.
It returns both the cleaned GSEA results and the summarized MP scores by sample.
}
\examples{
mp_score_results <- mp_score_data(gsea_final)
gsea_res_clean <- mp_score_results$gsea_res_clean
score_sum_by_cluster <- mp_score_results$score_sum_by_cluster
write.csv(gsea_res_clean, "gseaFinal_mpScore.csv", row.names = FALSE)
}
