% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mp_score_plot.R
\name{mp_score_plot}
\alias{mp_score_plot}
\title{Generate an MP Score Bar Plot}
\usage{
mp_score_plot(score_sum_by_cluster)
}
\arguments{
\item{score_sum_by_cluster}{A data frame summarizing the MP scores for each sample.}
}
\value{
A ggplot object that can be plotted or saved.
}
\description{
This function generates a bar plot to visualize the MP scores calculated for each sample.
}
\examples{
pdf("mp_score.pdf")
print(mp_score_plot(score_sum_by_cluster))
dev.off()
}
