% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGSEA_individual.R
\name{runGSEA_individual}
\alias{runGSEA_individual}
\title{Run GSEA on Individual Samples}
\usage{
runGSEA_individual(DEG, org, signif)
}
\arguments{
\item{DEG}{A data frame of differentially expressed genes (DEGs), with columns such as \code{sample}, \code{symbol}, \code{log2FC}, and a significance column.}

\item{org}{A character string indicating the organism, e.g., \code{"mouse"} or \code{"human"}. This will determine which gene sets are used for enrichment.}

\item{signif}{A character string representing the column name in the \code{DEG} data that contains the significance values (e.g., \code{"p_val"},\code{"p_val_adj"},\code{"non-selection"}).}
}
\value{
A list of GSEA plots, one for each sample.
}
\description{
This function performs Gene Set Enrichment Analysis (GSEA) for each individual sample provided in the \code{DEG} data frame.
It returns a list of GSEA plots, one for each sample.
}
\examples{
DEG <- read.csv('test-degdata.csv')
gsea_results <- runGSEA_individual(DEG, "mouse", "p_val_adj")
}
